Admin user created in migration but does not enter data in isAdmin --- created function changePassword().

+++ app/Http/Controllers/ClientController.php
@@ -14,9 +14,7 @@ class ClientController extends Controller
     function getAllClients()
     {
         log::info("estoy");
-        $clients = DB::select('select * from clients');
-
-        return $clients;
+        return DB::select('select * from clients');
     }
 
     function getClient(Request $request)
@@ -36,9 +34,13 @@ class ClientController extends Controller
 
     }
 
-
     function newClient(Request $request)                                // create and actualize a client
     {
+        $admin = DB::select('select isAdmin from users where id ="' . Auth::id() . '"');
+
+        if ($admin[0]->isAdmin === 0) {
+            return 'You do not have Administrator permissions';
+        }
 
         $validatedData = $request->validate([                           // validate the data format
             'name' => 'required|string|max:255',
@@ -48,6 +50,8 @@ class ClientController extends Controller
         ]);
 
         $cif = $request->input('cif');
+
+
         $name = $request->input('name');
         $surname = $request->input('surname');
         //$image = $request->input('image');
@@ -82,8 +86,18 @@ class ClientController extends Controller
 
     function updateClient(Request $request)
     {
-        log::info('estoy en updateclient');
-        log::info($request);
+        $admin = DB::select('select isAdmin from users where id ="' . Auth::id() . '"');
+
+        if ($admin[0]->isAdmin === 0) {
+            return 'You do not have Administrator permissions';
+        }
+
+        $validatedData = $request->validate([                           // validate the data format
+            'name' => 'required|string|max:255',
+            'surname' => 'required|string|max:255',
+            'cif' => 'required|string|max:255',
+        ]);
+
 
         $cif = $request->input('cif');
         $name = $request->input('name');
@@ -102,8 +116,7 @@ class ClientController extends Controller
             DB::select('update clients set name ="' . $name . '", mCIdUser ="' . Auth::id() . '"where cif="' . $cif . '"');
             return "Update Client CIF: " . $cif . " new name: " . $name;
         }
-        if (isset($surname))
-        {
+        if (isset($surname)) {
             log::info('estoy en surname');
             DB::select('update clients set surname ="' . $surname . '", mCIdUser ="' . Auth::id() . '"where cif="' . $cif . '"');
             return "Update Client CIF: " . $cif . " new surname: " . $surname;
@@ -113,14 +126,20 @@ class ClientController extends Controller
 
     function deleteClient(Request $request)
     {
+        $admin = DB::select('select isAdmin from users where id ="' . Auth::id() . '"');
+
+        if ($admin[0]->isAdmin === 0) {
+            return 'You do not have Administrator permissions';
+        }
+
         $validatedData = $request->validate([                           // validate the data format
             'cif' => 'required|string|max:255',
         ]);
+
         $cif = $request->input('cif');
         $client = DB::select('select * from clients where  cif ="' . $cif . '"');
-        if (isset($client))
-        {
-            DB::select('delete from clients where cif ="'.$cif.'"');
+        if (isset($client)) {
+            DB::select('delete from clients where cif ="' . $cif . '"');
             return "the user has been deleted";
         }
 
@@ -128,6 +147,12 @@ class ClientController extends Controller
 
     function updateImage(Request $request)
     {
+        $admin = DB::select('select isAdmin from users where id ="' . Auth::id() . '"');
+
+        if ($admin[0]->isAdmin === 0) {
+            return 'You do not have Administrator permissions';
+        }
+
         $validatedData = $request->validate([                           // validate the data format
             'cif' => 'required|string|max:255',
             'image' => 'required|image|dimensions:min_width=200,min_height=200',
@@ -137,8 +162,7 @@ class ClientController extends Controller
         $client = DB::select('select * from clients where  cif ="' . $cif . '"');
         log::info($client);
 
-        if (isset($client))
-        {
+        if (isset($client)) {
             $path = $validatedData['image']->store('public/storage');      // save image in images
             return DB::select('update clients set image ="' . $path . '", mCIdUser ="' . Auth::id() . '"where cif="' . $cif . '"');
         }
